# packages/dashboard/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ccusage-worv Dashboard API",
    "version": "1.0.0",
    "description": "Team usage collection dashboard API"
  },
...
  "paths": {
    "/api/usage": {
      "post": {
        "summary": "Submit usage report",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "memberName": {
                    "type": "string"
                  },
                  "sessionId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 512
                  },
                  "records": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "model": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 256
                        },
                        "inputTokens": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 100000000
                        },
                        "outputTokens": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 100000000
                        },
                        "cacheCreationTokens": {
                          "type": "number",
                          "default": 0
                        },
                        "cacheReadTokens": {
                          "type": "number",
                          "default": 0
                        },
                        "costUsd": {
                          "type": "number"
                        },
                        "projectName": {
                          "type": "string",
                          "maxLength": 256,
                          "default": ""
                        },
                        "recordedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "model",
                        "inputTokens",
                        "outputTokens",
                        "recordedAt"
                      ]
                    },
                    "minItems": 1,
                    "maxItems": 100
                  },
                  "reportedAt": {
                    "type": "string"
                  },
                  "toolUsage": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "toolName": {
                          "type": "string",
                          "minLength": 1
                        },
                        "callCount": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "acceptCount": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "rejectCount": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": [
                        "toolName",
                        "callCount",
                        "acceptCount",
                        "rejectCount"
                      ]
                    }
                  },
                  "turnCount": {
                    "type": "number"
                  },
                  "utilization": {
                    "type": "object",
                    "properties": {
                      "fiveHour": {
                        "type": [
                          "number",
                          "null"
                        ]
                      },
                      "sevenDay": {
                        "type": [
                          "number",
                          "null"
                        ]
                      },
                      "fiveHourResetsAt": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "sevenDayResetsAt": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      }
                    },
                    "required": [
                      "fiveHour",
                      "sevenDay"
                    ]
                  },
                  "pluginVersion": {
                    "type": "string",
                    "maxLength": 32
                  }
                },
                "required": [
                  "sessionId",
                  "records"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}